Add fields to detect profile field presence
Previously, any field not set by the server would have its corresponding
value in User set to its zero value. This is probably okay: an empty
string as a photo URL or username is almost certainly invalid.

Even so, it's nice to be able to know for sure which fields were omitted
and which ones had an explicit zero value. To accomplish this, this
change adds Has* fields to User when the distinction is important.